Fix grammar
" RTOS' " is incorrect. The apostrophe used at the end denotes that something belongs to RTOS (and would anyway be written RTOS's, unless RTOS was plural), whereas RTOS is an acronym or an initialism, depending on your pronunciation, and so the correct way to write 'more than one RTOS' is RTOSs.